031 - Update Manage structure of project
