+++ backend/src/main/java/com/example/backend/controllers/BrandController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-//@CrossOrigin(origins = "${app.cors.allowedOrigins}")
+
 @CrossOrigin(origins= "*")
 @RestController
 @RequestMapping("/itb-mshop/v1")
