+++ backend/src/main/java/com/example/backend/controllers/SaleItemController.java
@@ -18,9 +18,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-//@CrossOrigin(origins = "${app.cors.allowedOrigins}")
 @CrossOrigin(origins= "*")
-
 @RestController
 @Getter
 @Setter
