+++ frontend/src/App.vue
@@ -1,18 +1,17 @@
 <script setup>
 import { RouterView } from "vue-router";
-import Navbar from "./components/UI/Navbar.vue";
-import Footer from "./components/UI/Footer.vue";
-
+import Navbar from "./components/layout/Navbar.vue";
+import Footer from "./components/layout/Footer.vue";
 </script>
 <template>
-  <div class="bg-[#ffffff]">
-    <Navbar />
-    <main class="pt-15">
-      <RouterView />
-    </main>
+	<div class="bg-[#ffffff]">
+		<Navbar />
+		<main class="pt-15">
+			<RouterView />
+		</main>
 
-    <Footer />
-  </div>
+		<Footer />
+	</div>
 </template>
 
 <style scoped></style>
