+++ frontend/src/components/layout/Footer.vue
