+++ frontend/src/components/layout/Navbar.vue
