+++ frontend/src/components/product/SaleItemCard.vue
