+++ frontend/src/components/promo/PromoBar.vue
