--- frontend/src/mocks/sale-items.json
@@ -1,212 +1,0 @@
-[
-  {
-    "saleItemId": 1,
-    "model": "Galaxy S21",
-    "brandId": 1,
-    "description": "Flagship Samsung phone",
-    "price": 999,
-    "ramGB": 8,
-    "screenSizeInch": 6.2,
-    "storageGb": 128,
-    "color": "Phantom Gray",
-    "quantity": 50,
-    "createdOn": "2024-02-01T09:00:00",
-    "updatedOn": "2024-02-05T15:00:00"
-  },
-  {
-    "saleItemId": 2,
-    "model": "iPhone 13",
-    "brandId": 2,
-    "description": "Apple's 2021 flagship",
-    "price": 1099,
-    "ramGB": 6,
-    "screenSizeInch": 6.1,
-    "storageGb": 256,
-    "color": "Midnight",
-    "quantity": 30,
-    "createdOn": "2024-02-10T10:00:00",
-    "updatedOn": "2024-02-15T12:00:00"
-  },
-  {
-    "saleItemId": 3,
-    "model": "Mi 11",
-    "brandId": 3,
-    "description": "Xiaomi high-end smartphone",
-    "price": 699,
-    "ramGB": 8,
-    "screenSizeInch": 6.81,
-    "storageGb": 128,
-    "color": "Horizon Blue",
-    "quantity": 70,
-    "createdOn": "2024-02-12T11:00:00",
-    "updatedOn": "2024-02-17T14:00:00"
-  },
-  {
-    "saleItemId": 4,
-    "model": "OnePlus 9",
-    "brandId": 4,
-    "description": "Performance phone from OnePlus",
-    "price": 729,
-    "ramGB": 12,
-    "screenSizeInch": 6.55,
-    "storageGb": 256,
-    "color": "Winter Mist",
-    "quantity": 40,
-    "createdOn": "2024-02-13T09:30:00",
-    "updatedOn": "2024-02-18T16:30:00"
-  },
-  {
-    "saleItemId": 5,
-    "model": "Xperia 1 III",
-    "brandId": 5,
-    "description": "Sony's professional-grade smartphone",
-    "price": 1299,
-    "ramGB": 12,
-    "screenSizeInch": 6.5,
-    "storageGb": 256,
-    "color": "Frosted Black",
-    "quantity": 20,
-    "createdOn": "2024-02-14T10:10:00",
-    "updatedOn": "2024-02-19T11:50:00"
-  },
-  {
-    "saleItemId": 6,
-    "model": "Galaxy A52",
-    "brandId": 1,
-    "description": "Affordable Samsung mid-ranger",
-    "price": 399,
-    "ramGB": 6,
-    "screenSizeInch": 6.5,
-    "storageGb": 128,
-    "color": "Awesome Black",
-    "quantity": 80,
-    "createdOn": "2024-03-01T08:00:00",
-    "updatedOn": "2024-03-05T10:00:00"
-  },
-  {
-    "saleItemId": 7,
-    "model": "iPhone SE (2022)",
-    "brandId": 2,
-    "description": "Compact and powerful",
-    "price": 429,
-    "ramGB": 4,
-    "screenSizeInch": 4.7,
-    "storageGb": 64,
-    "color": "Product Red",
-    "quantity": 60,
-    "createdOn": "2024-03-02T09:15:00",
-    "updatedOn": "2024-03-06T11:30:00"
-  },
-  {
-    "saleItemId": 8,
-    "model": "Redmi Note 11",
-    "brandId": 3,
-    "description": "Great value for money",
-    "price": 199,
-    "ramGB": 4,
-    "screenSizeInch": 6.43,
-    "storageGb": 128,
-    "color": "Graphite Gray",
-    "quantity": 120,
-    "createdOn": "2024-03-03T07:45:00",
-    "updatedOn": "2024-03-07T09:00:00"
-  },
-  {
-    "saleItemId": 9,
-    "model": "OnePlus Nord CE 2",
-    "brandId": 4,
-    "description": "Balanced specs at a great price",
-    "price": 299,
-    "ramGB": 8,
-    "screenSizeInch": 6.43,
-    "storageGb": 128,
-    "color": "Bahama Blue",
-    "quantity": 100,
-    "createdOn": "2024-03-04T12:20:00",
-    "updatedOn": "2024-03-08T14:45:00"
-  },
-  {
-    "saleItemId": 10,
-    "model": "Xperia 10 III",
-    "brandId": 5,
-    "description": "Compact mid-range phone from Sony",
-    "price": 429,
-    "ramGB": 6,
-    "screenSizeInch": 6.0,
-    "storageGb": 128,
-    "color": "Blue",
-    "quantity": 35,
-    "createdOn": "2024-03-05T10:30:00",
-    "updatedOn": "2024-03-09T11:45:00"
-  },
-  {
-    "saleItemId": 11,
-    "model": "Galaxy Z Flip3",
-    "brandId": 1,
-    "description": "Foldable phone innovation",
-    "price": 999,
-    "ramGB": 8,
-    "screenSizeInch": 6.7,
-    "storageGb": 128,
-    "color": "Cream",
-    "quantity": 25,
-    "createdOn": "2024-03-06T10:00:00",
-    "updatedOn": "2024-03-10T10:30:00"
-  },
-  {
-    "saleItemId": 12,
-    "model": "iPhone 14 Pro",
-    "brandId": 2,
-    "description": "Top-tier Apple flagship",
-    "price": 1299,
-    "ramGB": 6,
-    "screenSizeInch": 6.1,
-    "storageGb": 512,
-    "color": "Deep Purple",
-    "quantity": 45,
-    "createdOn": "2024-03-07T11:10:00",
-    "updatedOn": "2024-03-11T12:00:00"
-  },
-  {
-    "saleItemId": 13,
-    "model": "POCO X3 Pro",
-    "brandId": 3,
-    "description": "Gaming phone at budget",
-    "price": 269,
-    "ramGB": 6,
-    "screenSizeInch": 6.67,
-    "storageGb": 128,
-    "color": "Phantom Black",
-    "quantity": 90,
-    "createdOn": "2024-03-08T13:00:00",
-    "updatedOn": "2024-03-12T14:00:00"
-  },
-  {
-    "saleItemId": 14,
-    "model": "OnePlus 10 Pro",
-    "brandId": 4,
-    "description": "Premium performance OnePlus",
-    "price": 899,
-    "ramGB": 12,
-    "screenSizeInch": 6.7,
-    "storageGb": 256,
-    "color": "Emerald Forest",
-    "quantity": 30,
-    "createdOn": "2024-03-09T15:20:00",
-    "updatedOn": "2024-03-13T16:00:00"
-  },
-  {
-    "saleItemId": 15,
-    "model": "Xperia Pro-I",
-    "brandId": 5,
-    "description": "Sony pro camera smartphone",
-    "price": 1799,
-    "ramGB": 12,
-    "screenSizeInch": 6.5,
-    "storageGb": 512,
-    "color": "Black",
-    "quantity": 10,
-    "createdOn": "2024-03-10T14:10:00",
-    "updatedOn": "2024-03-14T15:30:00"
-  }
-]
