--- frontend/src/views/LandingPageView.vue
@@ -1,11 +1,0 @@
-<script setup>
-import HeroBanner from "@/components/Home/HeroBanner.vue";
-import PromoBanner from "@/components/Home/PromoBanner.vue";
-</script>
-
-<template>
-  <PromoBanner />
-  <HeroBanner />
-</template>
-
-<style scoped></style>
