+++ frontend/src/views/error/Error500View.vue
