+++ frontend/src/views/home/LandingPageView.vue
@@ -1,0 +1,11 @@
+<script setup>
+import HeroBanner from "@/components/home/HeroBanner.vue";
+import PromoBanner from "@/components/promo/PromoBanner.vue";
+</script>
+
+<template>
+	<PromoBanner />
+	<HeroBanner />
+</template>
+
+<style scoped></style>
